Fix typeinfo not found
When building with the vptr sanitiser enabled, linking would fail with
messages about undefined references to Version_1_Nation.

This was due to the key function for Version_1_Nation (and
Version_1_NationBase) being undefined.

Since the dummy() function is never actually anyway (which would have
resulted in an error earlier on anyway), and it does not look like these
classes were meant to remain abstract, as at least Version_1_Nation is
instantiated, we can remove the dummy() function declarations.