World: Fix disappearing builder bug
It was possible to put a builder onto a tile that is walkable but not
buildable and then that tile would count as buildable when placing a
building, when normally it would not.  This is most evident on coastal
corners or near rocks, which have unbuildable tiles near them.  Issuing
the build order in this state would cause the unit to disappear (off the
map only) and become unusable while still counting as a unit for all
other purposes.

This was because the code to check whether a tile is buildable only
checks whether the building unit is the one blocking a tile, without
checking whether that tile would be otherwise buildable.

Fix this by also checking whether power is on for the tile, which is the
attribute that controls buildability on otherwise unoccupied tiles.  We
do not need to check that the tile is walkable because the builder unit
being there confirms that.

Ideally, this could all use a refactor.