Fix building menus bouncing to main on production start
The Town, Harbour, and War Factory production, spy, etc. menus were
bouncing back to the main menu when production started on a new unit.

This was due to the fix for production queue counts bouncing
(49faa54bec22eb0a1b78c3def6adaa83aaa9bf84) inadvertently breaking this
and causing the building to reset to the main menu any time production
started.  This is because info.disp() always resets the info pane back
to the main menu.  It is unclear why it was coded that way.

Fix this by only calling info.disp() (to refresh the production queue
count) if the building is already on its main menu.