Fix gettext version mismatch
Depending on the installed version of gettext, builds may fail with a
message similar to:

    *** error: gettext infrastructure mismatch: using a Makefile.in.in
    from gettext version 0.20 but the autoconf macros are from gettext
    version 0.22

This is because the macro that was being
used—AM_GNU_GETTEXT_VERSION—locks gettext to a specific version (or at
least it _should_, but it still allows *some* range of versions, it
seems.  It is not very clear at all how this works).

Fix this by instead using the AM_GNU_GETTEXT_REQUIRE_VERSION macro,
which requests gettext with a minimum version, allowing any version of
gettext from 0.19 onwards.