Fix mingw build with GCC 14
The mingw build using GCC 14 was failing with errors about mutexes,
threads, etc. being undefined.

This is because GCC 13 rewrote its threading and system code to work
better with the C++ standard library.  However, this means that older
versions of Windows are no longer supported.

Fix this by setting `_WIN32_WINNT` to `0x0600` (Vista) for the Windows
build.  This means that there is no longer support for older versions of
Windows.