Fix automatic caravan replacement segfault
Following an earthquake, if there were idle caravans which got
damaged (as all units outside of buildings do get), the game would crash
with a segfault.

This was because the automatic caravan replacement was trying to both
clone and retire the caravan at the same time, and so the route copying
code would try to read from a Unit that had already been deleted, for
which unit_array[] returns null, resulting in a segfault.

Fix this by preventing idle caravans from being cloned in
Polity::cloneCaravan().  Idle caravans are already prevented from being
cloned manually, but Polity::replaceCaravan() was below those checks.