Fix use-after-free after dismissing a unit
Dismissing (or disbanding or resigning) a unit would cause a memory
violation with a use-after-free.

This was due to ::Unit::detect_info() calling
::Unit::detect_select_hotkey() after the Unit was resigned and had
delete called on it.  This is a use-after-free memory violation and
undefined behaviour.

Fix this by moving the detect_select_hotkey() call to before the call to
group_resign()—and so the delete/free—can happen.