Fix load game screen file information legibility and position
The file name and file date on the save and load game was not very
legible and the positioning was inconsistent.

This was due to ::Font::put_to_buffer() acting differently to put() by
neglecting the inter-character space, and was revealed by the recent
change to UserInterface::printText() for clipping, which switched from
using put() to put_to_buffer().

Fix this by having put_to_buffer() account for the inter-character space
the same as put() does.