Fix build Fort keybind not working
When trying to press B-F to build a Fort, the game was instead selecting
a Fort.

This is because the keybind for selecting a Fort (F) was conflicting
with the select a Fort keybind.

Fix this by changing the default keybind for selecting a Fort to `N`.
Users can set their own keybind in the config.txt file if they wish.