Fix cloned caravans sometimes being inaccessible
Sometimes, cloning a caravan would result in the caravan being created
on another island or otherwise inaccessible location and not being able
to follow its route.

This was due to Polity::findNearest7kaaMarket() looking only at the
as-the-crow-flies distance and not taking into account
accessibility (which is tracked by 7kaa's regions).

Fix this by confirming that the Market being looked at is in the same
region as the location being compared to.