Fix shown target resistance when workers are present
Independent Towns were showing their target resistance incorrectly when
there were any workers in them.  Regardless of the Nation, the viewing
player would always see the resistance dropping to 0, even though this
is not actually happening.  The resistance for all races (and the
average) would be shown as dropping to 0, even when they did not have
any workers.

This is because the condition was previously mistakenly just whether the
Town had workers.

Fix this by correctly checking connected Firms for Workers for the
Nation.

Now, each race will show that it's dropping to 0 only if it has workers
for the Nation, and the average target resistance will correctly
calculate using the races dropping to zero and not.