ConfigAdv: Change mp_service_addr type to std::string.
We gain the use of the string copy constructor automatically, which allows
deep copying ConfigAdv to just work. mp_service_addr does not need to be
serialized, so no serialization methods are being provided for std::string.

Double free was possible previously if copying occurred. This was reported
by Infiltrator.