Fix not being able to immediately play replays
After a multiplayer game, pressing R to watch the replay was not working
or was loading an old replay.

This was because the replays were getting saved to the wrong
location (config directory instead of the save/multiplayer) directory,
but being read from the correct location.  This was done incorrectly in
the initial implementation of the directory split.