Fix compilation running even when unnecessary
When running `make`, Ambition_version.o, and so libambition.a and
7k-ambition, were always being recompiled, even when there were no
changes at all.

This is because of how git_string.h was defined in the Makefile, as a
.PHONY, meaning that it always counted as "changed".  Secondarily,
it *was* actually always being changed, because it was always being
written to.

Fix this by writing the git string to a temporary file, comparing, and
writing to git_string.h only if there is a change.  This way, we now
again get the holy "Nothing to be done" message.