Work around requested builders going to rally point instead
In multiplayer (but not singleplayer) builders which were requested from
buildings with rally points would start going to the requested building,
but then go to the rally points instead.

This was due to delay inherent in 7kaa's multiplayer between an order
being issued and executed.  This means that the builder is ordered to
leave the building, is ordered to go to the new building, actually
leaves, is ordered to go to the rally point, actually begins going to
the requested building, but then actually begins going to the rally
point.  This is not an issue in singleplayer because there is no delay
between the order being issued and executed.

Work around this by having builders ignore rally points.  This affects
more than quest request builder, as all builders will always ignore all
rally points, but until the issue-confirm-retry system is implemented,
this is the better compromise than request builder being unreliable.