Battle: Allow generating more Fryhtan lairs
The number of Fryhtan lairs generated on the map was limited to the
number of independent Towns generated, because they are generated in the
same loop, but the loop broke as soon as no more independent Towns could
be created (because of space constrains), so no more Fryhtan lairs could
be generated either.

Simply remove this `break` statement, to allow the loop to continue on,
attempting to create the requested number of Fryhtan lairs.