Fix replay desync on automatic caravan replacement
Replays were getting desyncs.

One cause of this was the checking of Caravans for automatic
replacement.  In replay mode, extra actions were being generated for
replacing the Caravan, as if the player were currently playing the game,
causing the replacement to happen earlier than what is recorded in the
replay.

Fix this by not checking Caravans for replacement in replay mode.

Incidentally, this should also fix the issue with multiple Ambition
clients all sending the same commands for replacing Caravans.