Fix segfault on hovering over caravan list with no caravans
Hovering over the Trade Report's caravan list (actually only where the
first caravan would be) when there are zero caravans was crashing.

This was due to using VBrowseIF::mouse_over(), which actually returns 1
when hovering over the area where the first record would be, even if
there are actually no records, and then this value is used to try to get
a record, resulting in a segfault.  This is due to VBrowseIF pretending
to have 1 record when it has 0.  It is not clear why this was coded this
way, especially since there is a separate `none_record` anyway.

Fix this by checking caravanBrowser.none_record and doing nothing and
returning early in the case that there are no caravans.  (This is what
the following code would have done, had mouse_over() returned the
correct value.)