Fix 7kaa config.txt settings not taking effect
When there was no Ambition config.txt file, the 7kaa config.txt settings
were not taking effect.

This was due to a leftover from the way 7kaa handles the config.txt file
where it discards all settings on any error (including the file not
existing).

Fix this by just removing that setting reset.