TownAI: Correct logic checking available product.
This fixes a regression introduced by 3d2d1e391 (2.14.5) which had fixed an
underflow. This accidentally excluded clay products from the town product
stats, which will have the effect of causing the AI to try to continually
import and export clay products. Fixing the logic error may reduce the AI
building markets, factories, or camels in cases where the supply was
present on site.

Regarding change 3d2d1e391, the underflow bug is also present in 7k2.

Thanks to DieselMachine for finding the bug.