Fix Encyclopaedia slideshow showing blank images
When the Encyclopaedia entered slideshow mode (by not moving the mouse
for 60s), all of the images were showing blank except for a brief flash
when switching entries, unless there were some user input.

This is because the next frame slideshow code puts a blank image onto
the front buffer and the actual image onto the back buffer, and then no
flip occurs and then Control::delayFrame() waits until the next time
that the entry should change.

Fix this by not running Control::delayFrame() immediately after the
entry switch for one frame, to allow a flip to occur.