Unit: Add option to disable space for attack check.
Before an attack is initiated, a space check around the target unit is
performed. After review of this code under test, no benefit really was
determined. Firstly, the unit will still need to travel to some location
if too far anyway. Second, if we take the check on face value, it is no
guarantee anyway during travel that the unit can reach the location --
it still have to find it's way there somehow.

But even then, there are further issues with what is being done. I
discovered this check because it was being performed with a unit that was
ranged. This check makes no sense for ranged units. And more, the check
is primarily intended to return true only if the blocked sides of a target
unit changes. It is not even a space check really. It's a change of space
and in code, the logic really reads "check for space". It seems like
almost nonsense, because this is done to decide to execute the action at
all, and executing the action is better to determine if it can be done
or not.

And another thing, why I was concerned and started digging into this, if
for some reason a unit executes a stop(), the blocked edges are not
cleared, and then can never be cleared, and when that happens, some, or
maybe all units that stop and are given an attack order after when it is
not cleared can never pass the check because if the blocked edges aren't
updates for whatever reason, this then will always cancel the attack
order. Note there is also a place in the code that just periodically
resets a unit's edge check randomly.

Testing shows that simply disabling this check may actually prevent bugs
at no loss really. The code isn't needed. 7k2 doesn't even have this code
btw. Time benchmarks show that there are no savings really in unit
processing by not doing this check, but actually a slight increase 0-2% of
additional CPU, maybe by enabling units that were previously stuck with
the order canceled prematurely.

I'm not sure if this is what was causing the other odd unit or pathing
behavior or one of the things.

Disable this for now, unless someone finds something that breaks then we
can revisit this.